requirements.txt
aws-cdk-lib==2.103.0
constructs>=10.0.0,<11.0.0
boto3>=1.28.57
awscli>=1.29.57
botocore>=1.31.57